Refactored to better align with Go language structure
Test cases are added
