+++ backend/internal/config/config.go
@@ -1,0 +1,6 @@
+package config
+
+const (
+	MongoURI = "mongodb://localhost:27017"
+	DBName   = "us_states"
+)
