+++ backend/internal/models/state.go
@@ -1,0 +1,6 @@
+package models
+
+type State struct {
+	Name string `json:"name" bson:"name"`
+	Code string `json:"code" bson:"code"`
+}
